Fix for llvm-3.7 ranlib on macports
macports fails now with the system ar -rcs, its
ranlib it cannot sort the symbols in the libponyc.a.
Hence check for llvm-ar-mp-3.7 and homebrew llvm-ar-3.7 with
a different mandatory AR_FLAGS=rcs
This is only needed for config=release, the -flto linker.

WIP: ld fails to find libLTO also. maybe need to add the macports
llvm-config --libpath.
See [ponyc #456]